smart-task-planner/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==1.10.13
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
httpx==0.27.0
python-dotenv==1.0.0